adds logging for parser and bridge metadata (#2192)
* adds logging for parser and bridge metadata

As part of adding more logging we refactor the logging of passes into
a Logging::PassManager, which is essentially a wrapper for the main
PassManager that overrides the init_apply and end_apply methors to
redirect logging. We replace the previous custom init_apply/end_apply
with inheritance from Logging::PassManager.

Of course, it can't be as simple, since parser logging is spread
through pass managers, inspectors, and functions. So we have added
an option to the FileLog object to append to an existing log.
With this support, we dump all parser logs in one file.

We publish the logs in the manifest file. With that come a set of
other changes, in particular related to handling run_id. Because it
uses the openssl library to compute the SHA, and thus ends up included
in gtests, we factor out the constructor and include the ssl library
only in a .cpp file. Now manifest.h can be safely included anywhere.

Finally, we make the `--verbose` option to the driver compatible with
Glass and take a verbosity argument. This is used mainly for PHV and
parser logging. Verbose 1 is LOG1, 2 is LOG3 or LOG 4 (depending on
module), and 3 is truly verbose -- 250MB of log file for a switch
compilation. And for good measure, we generate two of these, one for
each allocation pass. The more passes, the more logs.

Multiple fixes revealed by adding tests:
- add missing includes
- logs should not be generated unless verbose is > 0
- logs file path in the manifest is relative to the output directory
- add at least one verbose test to the option testing suite
- update the options in the test script (archive with name and verbose)