Gateway comparison fields must be sliced in the same way (#2203)
Ensure that fields involved in gateway comparisons are sliced in the same manner. Suppose we have two fields hdr1.f1 and hdr2.f2 that are used in the gateway expression if (hdr1.f1 == hdr.f2). This requires hdr1.f1 and hdr2.f2 to be aligned in their respective containers. However, the Clustering passes assume that both these fields are sliced in the same manner, without actually slicing fields so. This pass looks at fields involved in gateway expressions and ensures that they are sliced in the same manner.

Note that we do not need to actually calculate which bits of the field are used for the gateway, because that information is already present in the form of slices created during the MakeSlices pass.

P4C-647 #bug-present #assign #assign #start-work #fixed #fixed
P4C-976 #assign #start-work #fixed Update XFails and cpplint #fixed