Link p4c-barefoot statically (#2208)
* Link statically

In order to distribute a single binary for 32/64 bit Linuxes (Linuxii?) ...
Linux distributions :)

Tested: build on Debian 8, ran on Ubuntu 16.04, 14.04, Mint 18, and CentOS 7.

Other fixes:
* fixes for Debian 8 in install_os_deps
* If we requested static boost libraries, check not only the version,
but also the availability of the .a library
* Install the version of p4c_src/main.py that defaults to tofino/tna.
* misc optimizations
* Build tests only if ENABLE_TESTING

Also bail out before invoking the serialization if there were errors
in the frontend. Otherwise we endup with an invalid program and report
this as a bug.