consistent table naming for visualization (#2217)
* consistent table naming for visualization

We need consistent names between tables published in context.json
and resources.json. So we remove the suffixes for all table types
except $action.

We still need to do something about atcam partitions, e.g.,
$atcam$lt0, which do not show up in resources.json at all.

At the same time, it seems that for P4_14, we rename tables to unique
names unnecessarily. In the frontend, UniqueNames invoked twice gives
us back the original names. In the midend that doesn't work, so we do
end up with table names in context.json and resources.json that have no
corresponent in the P4. So we remove the renaming pass.

* fix hardcoded names in gtests

We hardcoded the table names based on the way the midend was renaming
tables. Since we removed the midend pass, now tables are no longer
suffixed by _<id>.

Also stumbled on a bug in the Clang implementation of handling
initializer_lists. When using -O3, the lists are completely mangled. I
guess they are just not supposed to be used as objects themselves,
because when we replaced them with a class that just took the
initializer_list as a constructor param, it works!

Also removed boost::is_initialized which has been
obsoleted. boost::none can be directly treated as a bool.

P4C-987 #assign #start-work #fixed consistent table naming for visualization #fixed