Proactively cut off slice lists with exact containers requirements. (#2223)
All field slices in a supercluster must be allocated to the same MAU group; thus, all these slices
must be allocated to containers of the same size. A supercluster may be composed of multiple slice
lists with the exact containers requirement. After preslicing, each of these slice lists with exact
container requirements must be of the same width; any other slicing generates unsatisfiable
constraints because slice lists with exact container requirements must be allocated to containers
with widths equal to the width of the slice list.

Consider `n` slices that belong to `n` different slice lists with exact containers requirement and
different widths, such that these `n` slices are all part of the same rotational cluster (i.e.  they
participate in MAU operations with each other). For generating allocatable superclusters, each of
these slice lists must be be sliced around the `n` slices to produce slice lists of the same width
(this width should be equal to the smallest width from among the `n` different slice lists).

In the existing `SlicingIterator` infrastructure, we ignore these MAU-induced constraints and
instead rely on the iterator to explore all possible slicings of the supercluster until we find one
that is allocatable. However, this often results in extremely long running compilation for large
superclusters (effectively an infinite loop). This PR instead accounts for these MAU-induced
constraints to generate an initial state of the `SlicingIterator` that is more likely to be
satisfiable, thereby pruning large parts of the search space explored by the `SlicingIterator`.

Note that this is not complete; there are still cases where preslicing might enter an "infinite"
loop. Subsequent PRs will further refine pre-slicing to eliminate those cases.