Mark intrinsic and invalidate from arch metadata (#2299)
user-accessible intrinsic metadata fields that are read at the deparser must be invalidated when
not written by the user (on Tofino). We introduce a separate property for the corresponding
fields--`invalidate_from_arch_i`--to indicate this. This frees up the `intrinsic` property to be
used for all intrinsic metadata.

Mark all intrinsic fields with the `intrinsic` property. These are fields extracted in the $entry_point, $ingress_metadata, and $egress_metadata parser states.

P4C-1066 #assign #start-work #fixed #fixed