add --ir-to-json option to driver and cleanup error codes
We add a --ir-to-json option to serialize the IR into a file (or
stdout).  This is the driver option that maps to the --toJson option
of the backend. The reason for a separate option is that we will
expose this option for an installed version of the compiler, while the
--toJson is exposed only in developer mode.

Also cleanup the error codes such that the driver can figure out
whether errors were user/program errors, compiler/fitting erorrs, or
compiler bugs. For fitting errors, we still want to generate a .bfa
and a context/resource file.

P4C-1032 #assign #start-work #fixed add --ir-to-json option to driver and cleanup error codes #fixed