Add support for ValueSet in TNA P4Info & BF-RT JSON
We now support PVS tables in the BF-RT JSON schema. Because the BF-RT
JSON is generated from the TNA P4Info message, P4Info needs to support
ValueSet as well. Even though ValueSet is a P4_16 built-in and is
supported by default in P4Info, that support proved not to be
sufficient. P4Runtime v1.0 restricts ValueSet instances to pack a single
binary string and does not support structs or tuples. Which is why we
had to introduce a TNA-specific ValueSet Protobuf message in
barefoot/p4info.proto. This message is generated by the Tofino P4Runtime
ArchHandler code and is then translated to BF-RT JSON.

This commit also includes a test program: pvs_tna.p4.

P4C-1083 #fixed #fixed