Live range shrinking pass enabled by metadata initialization (#2181)
* Remove old metadata initialization pass.
* Add the MapTablesToActions class. Generate O(1) lookup maps for tables to set of actions that can
  be invoked by them and the set of default actions for each table.
* Add FindInitializationNode class. Call to FindInitializationNode::findInitializationNodes
  generates the set of actions where the queried fields must be initialized for the queried metadata
  fields to share the same container.
* Integrate metadata initialization with PHV allocation
* Add AddMetadataInitialization pass manager
* Change AllocatePHV heuristics. Also, disable overlay across gresses.
* Add ability to overlap with fields that are only read in the last used stage.
* Fix the validation check
* Initialization information is now carried forward to alloc_slice objects
* Fix a problem for when a field is only used in the parser.
* Remove FindInitializableFields pass.
* Add initialization for metadata that show overlays due to live range shrinking.
* Add initialization for marked fields.
* Push initialization down from group dominator
* Penalize new containers the most; more than overlay bits
* satisfies_constraints now takes into account metadata initialization. We prevent the packing
  together of fields that are extracted and also have an uninitialized read.  However, this needs to
  account for the fact that fields initialized with metadata initialization no longer have
  uninitialized reads.
* MetadataInitialization -> LiveRangeShrinking
* Disable overly conservative checks for init points
* Fix the initialization of fields
* Account for deparse zero overlays
* Introduce map of metadata initialized in Allocation objects.

* Fix initialization problem
I was previously looking at action names before initializing, which was causing spurious
initializations to be inserted in other actions with the same name. Now, I use action pointers to
only initialize at the right action.

* Special handling when deparser is not dominated by a non gateway table.
* Account for metadata fields that may be overlayed because of mutual exclusion. Only consider one of these fields while figuring out initialization points.
* Some logging changes
* Convert overly restrictive checks for hash action to warnings
* Change DEP_DIST to 3
* Include tbl_act in analysis

* Determine what dependencies need to be enforced after live range shrinking
We need to make sure that the live range assumptions after live range shrinking still hold valid
after the metadata fields are initialized. For this purpose, we should enforce dependencies between
the usedefs of the metadata field and the initialization points for the metadata field it overlays
with (and has a later live range).

* Add dependency based on live range shrinking assumptions
* Fix error related to what was considered by AllocatePHV as uninitialized
* When group dominator stage < allowedStage, then initialization is not possible. Also, update XFails and cpplint and some code refactoring
* BASIC master fits for this commit. A bunch of knobs turned for fitting. Test with other profiles.

* Prevent extension of the dependence chain for the last three stages
Live range shrinking may introduce new dependencies between tables. For the case where we initialize
at multiple strict dominators (instead of the group dominator), new anti and output dependencies
would be introduced between each strict dominator where initialization happens. This commit
introduces a heuristic where we avoid initializing at strict dominators if any of the strict
dominators have min stage in one of the last three stages, determined by the dependence graph. This
is a coarse check that prevents the extension of the dependence tail.

* Prevent metadata initialization from extending dependency critical path
Initialization at a group dominator may extend the dependency chain if a use of the field being
initialized and the group dominator have the same min_stage value. Recognize these cases, and in
that case, initialize at a dominator node higher up the chain to the source.

* XFail updates and code cleanup for review

* Use backtracking for case where meta init results in >Device::numStages table placement
Because of the absence of unification of dependencies in the table dependency graph and better
interfaces from the dependency graph to metadata initialization, we may get table placements greater
than Device::numStages(), even when container conflicts are disabled. For such cases, we redo PHV
allocation without metadata initialization. This leads to us fitting switch p4_16.

* Address review comments

P4C-1037 #start-work #assign #assign #fixed #fixed