Bridged Metadata Packing Improvements and Metadata Slice List Slicing (#2412)
* Improvements to bridged metadata packing

When the following conditions are all true:
(1) Bridged fields are used as sources in the same actions;
(2) Their respective destinations are in the same byte; and
(3) The number of bridged fields is >2, or there are other bits within the destination byte that are
not written to in that action;
the bridged metadata fields must all be packed together in the same container, and their relative
alignment must reflect the alignment of the destination byte.

We detect this case and introduce a new must-pack constraint that takes precedence over all other
packing constraints during bridged metadata packing. The lack of this feature is causing PHV
allocation for one supercluster in an Arista profile to fail.

Additionally, we used to pack bridged fields that are not packed with any other field by default in the lowest
bits of the container. However, if the field has a different alignment requirement, and is used in
an action that requires PHV alignment (for e.g. other fields in the destination are written using
action data/constant value), then this bridged packing would cause PHV allocation to fail.

In this commit, I introduce a lightweight analysis that determines if the alignment constraint must
be respected for the bridged field, and if yes, pads this field accordingly.

P4C-1138 #bug-present #start-work #assign #assign #fixed #fixed

* Improvements to slicing
In the earlier slicing PR (#2400), we only sliced slice lists with exact containers requirements to fit the MAU constraints. However, we also need to slice metadata slices that participate in MAU operations with the aforementioned clusters with exact containers requirements. This patch introduces slicing for such metadata fields.