No-split for pa_container_size specification (#2524)
When a field is marked as pa_container_size and a single container greater than the field size is
specified, the programmer assumption is that the entire field will be allocated to the same
container. Previously, Brig could have allocated different slices of field to different
specified-size containers. Now, by adding a no-split constraint to the field in question, we ensure
that the entire field is allocated to the same specified-size container.