Support for Snapshot in P4Info and BFRT JSON (#2516)
* Support for Snapshot in P4Info and BFRT JSON

This commit adds support to generate Snapshot messages in P4Info
containing the necessary information for generating the schema for the
SNAPSHOT_<GRESS> and SNAPSHOT_<GRESS>_LIVENESS tables in the BFRT
JSON. One major change is that we had to include the notion of
"container" in the schema to accomodate for the control information
provided by the snapshot registers: identical information (e.g. next
table pointer, match hit address) is repeated for each stage and
sometimes for each logical table in a stage. We use repeated containers
(one per stage / one per logical table) to represent this
information. When parsing the information returned by the BFRT API, the
client will first retrieve the container by name / id (as it would for a
singleton field), the call the "Get" API again to retrieve the
appropriate singleton field(s) by name in the returned container. This
was agreed upon with Dileep.

P4C-1207 #fixed #fixed

* Incorporated review comments from BFRT team