Add check for constant extractors to POVs (#2540)
This PR gathers all the POV fields that are written using constant extracts in the same parser state.
The information about these fields is stored in newly introduced members of the PhvInfo class. A
new class, called TofinoParserConstantExtract walks through the parser state and populates these
members. As this constraint is only applicable to Tofino, the information is only gathered when the
compilation target is Tofino.

If a packing of POV bits packs results in two POV fields written using constant extractors in the
same parser state, then the packing must respect the following constraints to ensure only one extractor is used:

For 16b containers, the fields should be within 4 consecutive bits.
For 8b containers, the fields should be within 3 consecutive bits.
However, the compiler currently ignores the above constraint for a single extractor and generates a parser that has one extractor for both fields, even when they are separated by more than the allowed number of bits. This commit adds a check to the can_pack() method of ActionPhvConstraints, which ensures that POV bits written using constant extractors in the same parser state subscribe to the constraints above.

A separate PR will fold this constraint into POV slice list creation to ensure the use of minimal number of constant extractors.