Miscellaneous Slicing Fixes (#2550)
* Refactor slicing iterator constructors into smaller methods: one for populating data structures that
   are later used by a second method to impose MAU constraints during slicing.
* Overhaul the way the intermediate information about slicing is propagated through the maps
   during slicing. This resolves a bunch of corner cases that were causing bad slicings to appear.

The basic idea behind the `impose_MAU_constraints` changes is that we identify the candidate
slice which necessitates a slice list to be divided into two. The slicing point could either be
anywhere to the left of this slice (up to the right limit of the last slice list that was created) or exactly
after the candidate slice (for cases where the candidate slice is byte aligned and satisfies the new
slice list size requirements). `sliceListToProcess` is the list of slices that need to be broken up
based on the identified slicing point.