Check that group dominator does not reach uses of fields with earlier live ranges (#2559)
* Fix bug with live range shrinking that was causing MSDC_IPV4_PROFILE to fail.

Add a missing check to ensure that the initialization point cannot reach the defuses of fields with
earlier live ranges.

Consider field `f` and a set of fields `G`, which are all overlayed because of live range shrinking.
Suppose all fields in `G` have earlier live ranges compared to the live range of `f`. Also, note
that the initialization points for the relevant fields in `G` have already been determined by the
time we are trying to determine the initialization point for `f`.

If `f` has an uninitialized read, the first candidate for the initialization point of `f` is the
group dominator for all its defuses. However, before going ahead with initialization, we must check
that the group dominator does not reach any use of the fields in `G`, i.e. that the group dominator
table does not have a dependency from itself to the tables in `G`. This check was missing in the
existing implementation of live range shrinking and was causing `MSDC_IPV4_PROFILE` to fail.

P4C-1272 #assign #start-work #fixed Fix bug with live range shrinking that was causing MSDC_IPV4_PROFILE to fail. #fixed

* Update XFails

P4C-1272 #assign #start-work #fixed Update XFails #fixed