fix segfault when erasing non-existing elements in a set
set.erase(begin, end) expects the element to be present in the
set. If not, it throws an exception. This fixes the issue by iterating
through the set, however, this actually points to a deeper issue -- it
seems that we need to find and eliminate those elements, but parser
inlining renames the header in such a way that the invariant no longer
holds.

P4C-1276