Support for sum-of constraint for slices of the same field (#2578)
When I change the dependence distance for live range shrinking to 2, switch DC_BASIC fails
because a 24b field requiring a hash action is split into three 8b slices and allocated across 3
different 16b containers. This allocation cannot be implemented in the compiler as there cannot be 
more than 4 immediate bytes per logical table and this allocation requires 6 immediate bytes.

The solution is to add a sum-of constraint to PHV allocation, which requires the sum of container
bytes for individual fields to not exceed a resource-dictated limit. This PR implements the case of a
sum-of constraint on slices of individual fields.

P4C-1278 #assign #start-work #fixed Add single field sum-of constraint. #fixed