Once metadata initialization is disabled, keep it disabled for rest of compilation
One of the possible reasons for the second round of table placement (the one without container
conflicts) requiring more than the maximum number of stages available, is the presence of metadata
initialization and live range shrinking. Therefore, in this case, we redo PHV allocation, with live
range shrinking and metadata initialization disabled and use that in subsequent PHV allocation and
table placement passes.

However, a bug in the implementation means that live range shrinking and metadata initialization are
enabled again, if a third round of PHV allocation is required by the program. This PR ensures that
live range analysis and metadata initialization, once disabled, is not enabled for the rest of the
compilation process. This fixes the fitting issue we intermittently keep seeing for p4-16 switch.