Implement constraint related to constant extractors (#2617)
* Implement constraint related to constant extractors

When we use a constant extractor to set a field in the parser, we need the constant to be
expressible within 3 bits when the field is allocated to a 32-bit container, and within 4-bits when
the field is allocated to a 16-bit container. This PR implements this constraint by checking if a
constant used as an extractor source can be expressed within 3 consecutive bits. If not, it adds a
constraint, requiring the field to go to 8b containers.

Note that we currently do not have a mechanism to prevent the allocation of a field to a certain
container size. Therefore, we do not distinguish between the 3b case for 32b containers and the 4b
case for 16b containers. When the `do-not-allocate-to-particular-container-size` constraint is
implemented, I will relax add this distinction.

P4C-1296 #assign #start-work #fixed Implement constraint related to constant extractors #fixed

* Add one negative test and one customer case that was showing up this bug.

P4C-1296 #assign #start-work #fixed Add one negative test and one customer case that was showing up this bug. #fixed

* Update XFail

Also, avoid the case where a stack validity bit greater in size than 8b will need to be
assigned to an 8b container. For such cases, we will rely on parser splitting states properly to
enabled the setting of metadata.

P4C-1296 #assign #start-work #fixed Update XFail #fixed