Add better diagnostics for PHV failures (#2629)
Add handle for diagnosing failures of unallocated superclusters
Introduce a diagnoseFailures() method that will check all unallocated superclusters and
determine if the unallocated superclusters can ever be allocated.

Introduce hook into ActionPhvConstraints for diagnosing allocation failure
The diagnoseSuperCluster() method in ActionPhvConstraints will be responsible for reporting
detailed error messages on PHV allocation failure.

Add diagnostics when an action requires too many sources
An action can only have two PHV and no action data/constant sources or one PHV source and one
action data/constant source. This PR adds diagnostics for this case, when slice lists in a
supercluster are at their minimal size (supercluster cannot be split further).

Add check for non-contiguous writes that cannot be realized by deposit-field
When some fields in the destination are not written in an action, one of the sources is the destination
itself. At least one of the set of slices written to or the slices not written to must be contiguous for
deposit-field to be used.