use p4c/working-top for switch.p4-16 regressions (#2624)
* use p4c/working-top for switch.p4-16 regressions
* switch.p4-16: resolve race condition with Tofino tests and xfail JBay

We were relying on the first test to compile switch and then we ran
tests in parallel. However, that introduces a race condition, so we
split the compilation job and make the tests depend on it.

Also, switch.p4-16 master fails for JBay.

* force binary generation even when there is no test specified

In the case where we want to pre-compile a P4 program and then
run tests, we want to generate a binary. Until now, runtest was
considering that if there are no tests specified (PTF, STF, PD, etc.)
there is no need to produce a binary. We add an option to runtest,
-force-link, to produce the binary.

Some other cosmetic cleanups, such as checking for the existence of
directories before they are removed, using $pipes only if the manifest
could be parsed, and so on. It does seem that runtest is in need of
some more housekeeping.

* Runtest refactor: P4C-856

* pipes extraction to be preceded by native compilation, update JBay Xfail

* Add tofino32q-3pipe program as compile only as multi-pipe PTF run is not supported by ptf_runner yet