Allow slices of fields to be allocated to TPHV (#2724)
Currently, PHV allocation forces all slices of a field to go into a PHV container, if even one of
those slices is used in the MAU pipeline. However, it should be possible for individual slices of
the field, which are only used in the parser/deparser but not in the MAU, to go into TPHVs. This PR
implements this change.