Account for pack conflicts during metadata initialization (#2735)
* Account for pack conflicts during metadata initialization

If initializing metadata at a table will induce a container conflict, do not initialize at that
table. This ensures that we do not lengthen the dependency chain and tables have a better shot at
fitting. This now results in MSDC_PROFILE fitting.

* Relax the pack conflict check

If initTable and table writing another field in the same container are mutually exclusive, then
initialization is still possible.