Remove a possible infinite loop within AllocatePHV (#2422)
When conditional constraints are generated corresponding to slices in a slice list, we only need to
check that the required alignment determined by the conditional constraints are satisfied by the
offset of slices within the parent slice list. If this check is true, then the required alignment
will be satisfied later during the allocation of the parent slice list later on during the
allocation of the same supercluster. Therefore, we avoid a whole bunch of unnecessary recursive
calls, which in some cases (see attached program), may even result in the manifestation of an
infinite loop.

P4C-713 #bug-present #assign #assign #start-work #fixed #fixed