Ability to force generation of "standard" P4Info for TNA program
One can now generate a standard P4Info message for a TNA program, which
means standard P4Runtime applications are now supported for TNA programs
running on Tofino.

To generate the standard P4Info message, use the
'--p4runtime-force-std-externs' flag in addition to '--p4runtime-file'
and '--p4runtime-format'.

The following features / externs are supported: [direct] meters,
[direct] counters, digests, action profiles / selectors and idle
timeout. If a TNA program includes a extern which is not supported
(e.g. LPF), a warning will be displayed.

I originally wanted to change the P4Info generation code directly, but
in the end it was much easier to just convert the Tofino-specific P4Info
message. This means that we can also generate the BF-RT JSON file as
before, which can be convenient (we can have the 2 services running
side-by-side on the switch, managing different objects).