Introduce a pass to determine potential overlays, assuming movement of fields into dark containers (#2789)
This class marks fields that could be put into the same container because one of them is moved into
a dark container. After this commit, `dark_live_range:1` prints the stages where each field in the
program is used (read or write) so that we can then decide overlays made possible by moving a field
into a dark container.