Remove DarkPrivatization pass (#2802)
This was a pass introduced to allow for some dark PHV usage; however, it is starting to affect PHV
on switch p4-16 profiles. This PR disables this pass, but keeps the implementation of the pass
around so that we may resurrect it, if needed, before the stage-based allocation is ready to merge.