staged Travis runs (#2817)
* staged Travis runs

Refactor the Travis jobs using stages. Stages are jobs that run
in sequence. Inside a stage, jobs run in parallel. Therefore, we
structure two stages:
 - one with a single job to wait for the Jenkins image build
 - one with 3 jobs that run the tests only when the image was built

With this setup, we minimize the number of concurrent jobs that wait,
and in case the image fails to build, we do not even start the 3 jobs.

* avoid cloning submodules

* Change build parallel level to j16

* Load balance Travis Tofino runs