Disable initialization for padding fields (#2853)
* Disable initialization for padding fields

Currently, we rely on live range shrinking pass to find slices that can be overlaid with
compiler-generated padding fields. The defining feature of these padding fields though is that they
can be used without initialization to 0, as we do not care about the values they contain. Therefore,
this PR removes the initialization that was earlier required for padding fields (marked by the
`alwaysPackable` constraint). This will potentially allow us more metadata overlay for all programs.

* Update XFails

* Change alwaysPackable -> overlayablePadding