Fix MSDC_SPINE_DTEL_INT_PROFILE compilation (#2857)
* Fix MSDC_SPINE_DTEL_INT_PROFILE compilation

In case deparsed-zero slices are overlaid with other metadata due to live range shrinking, it
results in the insertion of initialization instructions for each byte-sized deparsed-zero slice.
However, these multiple initializations of the same container will result in an error related to the
multiple use of the same instruction slot.

Therefore, for these cases, we need to check the allocation of slices that must be initialized in
the same action, and deduplicate the set of slices to be initialized, based on the container slices
allocated to each of those slices.

* Add test. Allow live range shrinking to happen for all fields in the program.

There is no reason to restrict live range shrinking to particular kinds of metadata. In case of
headers, there will be no disjoint live range, which will prevent overlay due to live range
shrinking anyway.