Fixes for aliasing and flexible packing (#2866)
This issue was exposed by the 32Q switch, where an alias destination expression was not found in the
second pipe and so, aliasing could not work properly. This fix involves changing the mechanism for
finding alias destination expressions to use the `gen_fieldref()` method. This method generates an
IR::Member object directly from the header object and the required field name.

This change, however, requires that all headers are flexible structs are flattened by the time we
implement aliasing, so I have had to lift the `FlexiblePacking` pass above the `Alias` pass.

P4C-1524 #assign #start-work #fixed Fixes for aliasing and flexible packing