Support for multi parser in BF-RT JSON generation
Because of the architecture definition, which supports providing
different parsers for each pipe and each gress, we have a different
BF-RT table for each (pipe, gress). If a pipe is not defined, then there
is no table.

To identify parser implementations, we are using the architecture name
("ingress_parserX" / "egress_parserX") as this is the only name which is
guaranteed to exist and be unique. An alternative would be to use the
user-provided name when available, but then the naming scheme would not
be consistent.

P4C-1556 #fixed #fixed