manifest schema 2.0 (#2895)
manifest schema 2.0

Reorganize the manifest schema and the compiler output directory per
pipe.  All the files (outputs, logs, graphs, binaries) are organized
output in their individual pipe directories, except for those files
that are global: manifest.json, the preprocessed source (name.p4pp),
bfrt.json, and the generated driver configuration file.

The manifest class itself has been rewritten to reflect the new
structure. In addition, the manifest is used as a visitor to generate
graphs. TODO: figure out why control graphs can't be generated
individually and need to be generated from the top level.

Refactored the handling of creating files and directory to rely on the
BFNContext::getOutputDirectory method, which takes a pipe argument and
returns the appropriately created directory. This removed a series of
repeated code that creates directories, and removed the need of a set
order of invocation for directory creation.

We also add support for driver config generation, by merging the
p4c-gen-bfrt-conf and gen_conf_pd.py and adding an argument to select
which configuration. TODO: deprecate and remove the older scripts --
however, this requires coordinating changes to p4-build, switch,
runtest, and jenkins scripts. For now it is just invoked in the
compiler driver and generates a conf file at the top level. All other
uses are generating config files after the compiler, so this should
not create an issue.

* manifest version 2.0.1
- lower case enums
- dump the compilation_time as string
- use packaging.version library to compare versions
- have the default python3 take precendence over /usr/local/bin/python3
- move .json files to the log directory

We decided that the .json files are logs, as reflected by the fact
that p4c-build-logs is generating them and thus we store them in the
logs directory. The manifest should find them in the right path.

Also renamed res.json to resources.json and cleaned up the registration
of the files in the manifest.