Inject Metadata Dependencies (#2923)
* Added a pass to inject CONTROL dependencies for metadata initialization

Added some comments and did some code cleanup

Added BUG in table placement which was the previous check for dependencies

Putting the inject dpendency directly on the two tables in order to correctly propagate chain lengths

Missing a clear in the flow graph.

* Fix issue with dependencies added due to metadata initialization

Even though the dependency information is calculated on a slice by slice basis, the
AddInitialization pass was converting this information from slice granularity to field granularity
before adding dependencies between the initialization tables and the use tables. This ended up in
false dependencies getting added, because of a loss of precision when going from slice to field
granularity. This PR fixes that issue.