Fix a couple of issues related to metadata overlay (#2844)
* Fix issue with packing incorrectly being treated as metadata overlay

When we check if multiple fields can be packed into the same container, we remove slices that are
overlayable through live range shrinking (metadata overlay) before calling
`ActionPhvConstraints::can_pack()`. However, the check was incorrect in that it checked whether
slices were potentially overlayable without checking if the slices were actually overlayed. This PR
fixes this bug.

* Add missing initialization leading to MalformedPacketsTest failure

Even when all strict dominators write to the candidate field for initialization, we also need to
ensure that all actions of the strict dominators have written to the field.

* Fix issue with `actions_writing_fields`

* Better handling of ImplicitParserInit

* Add STF test for the XOR packing vs overlay scenario

* Change the way disabling metadata initialization is implemented.

When we have specified that metadata initialization be disabled (happens when we get a container
conflict free table placement requiring more than the `Device::numStages()` number of stages), we
only disable metadata overlay that requires initialization instructions to be inserted. If
initialization instructions are not inserted, then there is no change to the table graph. This also
allows us to allocated fields to overlap with compiler-inserted padding fields (we treat allocation
over padding fields as a live range shrinking based overlay).

* Clean up the logging levels for LiveRangeShrinking related passes

* XFail updates

* Update fabric refpoint

* Add log message

* Add test for strict dominators writing only in some actions

Also, more accurate check for ImplicitParserInit in LiveRangeShrinking

* Fix an unhandle ActionPhvConstraints case

If the source container is not aligned with the container state being written, then the written
container state must be `src1`, as it would be shifted into the right bit position within the
container. This requires the written container state to be contiguous as the mask is expressed as a
contiguous [`lo`, `hi`] value for the associated `deposit-field` instruction. Therefore, for the
unaligned case, we should only check mask validity for the written container state.

P4C-1512 #bug-present #assign #start-work #fixed #fixed

* Remove MSDC IPV4 8.7 from XFails

* Update XFails

XFail the tests that are now taking more than 12 stages.

* Update XFails