ubuntu 18.04 (#2956)
* fixes (warning removal mainly) for Ubuntu 18.04

* Prevent inlining of IdletimeTable::write_merge_regs

- gcc7 somehow inlines this into compilation units that don't have
  access to the template it calls (link time optimization?  polymorphic
  specialization?) which then causes a link error.

* add skeleton for testing compiling the compiler on different Ubuntu variants.

Adds a set of Docker files that install bf-p4c dependencies on a
barebones Ubuntu variant (16.04 and 18.04 for now. 14.04 is a
mess!). With these, we can deploy jobs that simply verify that the
compiler compiles. No testing necessary at this time.

TODO: resolve how we install bf-utils. Maybe another container on top
of the base image?

* update test failures due to new p4c tests